Experiment with log transform
To avvoid having negative prediction we can use log transform on
the input data and then use the exponential to return the correct
prediction. This will give a lower bound of zero on the output.

One bug with the current solution with only using the log on the input
and exp on the output is a rather large noise element on the diagonal in
the covariance matrix. Giving a result where the noise is dominating the
variance.