New GP Taylor approximation
The GP now uses a fist order Taylor approximation,
calculated with the casadi jacobian of the GP mean.

Both the normal GP, and the Taylor approximation
is pre-built in the GP constructor, and stored as
private functions in the GP object.

This reduces the computation time for the TA from >1ms
to about 270us on a model with Nx=6 and N=50.